+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased]
 
+### Fixed
+
+ - Fix yaml syntax for Contao 4.9 compatibility
+
 ## [2.1.1] - 2020-02-07
 
 ### Fixed